requirements.txt
click==8.1.7
dnspython==2.6.1
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
rich==13.7.1
shellingham==1.5.4
typer==0.12.3
typing_extensions==4.11.0
